requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.0
mcp==0.9.0
tiktoken==0.6.0
